+++ packages/elements/modal-dialog/src/components/Modal.js
@@ -46,6 +46,7 @@ function getInitialState() {
   return {
     dialogNode: null,
     scrollDistance: getScrollDistance(),
+    hasEntered: false,
     isExiting: false,
   };
 }
@@ -103,7 +104,7 @@ type Props = {
   /**
     Function that will be called when the enter transition is complete.
   */
-  onOpenComplete?: ElementType => void,
+  onOpenComplete?: (node: ElementType, isAppearing: boolean) => void,
   /**
     Function that will be called when the modal changes position in the stack.
   */
@@ -142,6 +143,7 @@ type State = {
   dialogNode: Node | null,
   scrollDistance: number,
   isExiting: boolean,
+  hasEntered: boolean,
 };
 
 class Modal extends Component<Props, State> {
@@ -184,6 +186,14 @@ class Modal extends Component<Props, State> {
   handleDialogClick = event => {
     event.stopPropagation();
   };
+  handleEntered = (...args) => {
+    this.setState({
+      hasEntered: true,
+    });
+    if (this.props.onOpenComplete) {
+      this.props.onOpenComplete(...args);
+    }
+  };
   handleExit = () => {
     window.removeEventListener('scroll', this.handleWindowScroll);
     // disable FocusLock *before* unmount. animation may end after a new modal
@@ -206,7 +216,6 @@ class Modal extends Component<Props, State> {
       isChromeless,
       onClose,
       onCloseComplete,
-      onOpenComplete,
       onStackChange,
       shouldCloseOnEscapePress,
       stackIndex,
@@ -225,6 +234,13 @@ class Modal extends Component<Props, State> {
     const widthName = WIDTH_ENUM.values.includes(width) ? width : null;
     const widthValue = widthName ? null : width;
 
+    // Pass an afterEnded custom transition to Positioner so we can update styles to remove the transform property
+    // This fixes an issue with react-beautiful-dnd within modals - AK-4328
+    const customTransition =
+      this.state.hasEntered && !this.state.isExiting && !isBackground
+        ? 'afterEntered'
+        : '';
+
     return (
       <FillScreen
         {...transitionProps}
@@ -235,8 +251,9 @@ class Modal extends Component<Props, State> {
         <Blanket isTinted onBlanketClicked={this.handleOverlayClick} />
         <Positioner
           {...transitionProps}
+          customTransition={customTransition}
           onClick={this.handleOverlayClick}
-          onEntered={onOpenComplete}
+          onEntered={this.handleEntered}
           onExited={onCloseComplete}
           scrollBehavior={scrollBehavior}
           widthName={widthName}