+++ packages/fabric/editor-core/__tests__/editor/plugins/placeholder-text.ts
@@ -7,6 +7,7 @@ import {
   showPlaceholderFloatingToolbar,
   hidePlaceholderFloatingToolbar,
 } from '../../../src/editor/plugins/placeholder-text/actions';
+import { FakeTextCursorSelection } from '../../../src/editor/plugins/fake-text-cursor/cursor';
 import {
   createEditor,
   doc,
@@ -88,7 +89,7 @@ describe(name, () => {
         );
         expect(dispatchSpy).toHaveBeenCalled();
         expect(dispatchSpy.mock.calls[0][0].getMeta(pluginKey)).toEqual({
-          showInsertPanelAt: false,
+          showInsertPanelAt: null,
         });
       });
     });
@@ -110,17 +111,36 @@ describe(name, () => {
         showPlaceholderFloatingToolbar(editorView.state, editorView.dispatch);
         expect(editorView.state.doc).toEqualDocument(doc(p('hel{<>}')));
       });
+
+      it('should show the Fake Text Cursor when inserting placeholder text ', () => {
+        const { editorView } = editor(doc(p('hello{<>}')));
+        showPlaceholderFloatingToolbar(editorView.state, editorView.dispatch);
+        expect(editorView.state.selection).toBeInstanceOf(
+          FakeTextCursorSelection,
+        );
+      });
     });
 
     describe('hidePlaceholderFloatingToolbar', () => {
-      it('should set the `showInsertPanelAt` meta value to false', () => {
+      it('should set the `showInsertPanelAt` meta value to null', () => {
         const { editorView } = editor(doc(p('hello{<>}')));
         const dispatchSpy = jest.spyOn(editorView, 'dispatch');
         hidePlaceholderFloatingToolbar(editorView.state, editorView.dispatch);
         expect(dispatchSpy).toHaveBeenCalled();
         expect(dispatchSpy.mock.calls[0][0].getMeta(pluginKey)).toEqual({
-          showInsertPanelAt: false,
+          showInsertPanelAt: null,
         });
+      });
+
+      it('should remove the show the Fake Text Cursor when inserting placeholder text ', () => {
+        const { editorView } = editor(doc(p('hello{<>}')));
+        showPlaceholderFloatingToolbar(editorView.state, editorView.dispatch);
+        expect(editorView.state.selection).toBeInstanceOf(
+          FakeTextCursorSelection,
+        );
+
+        hidePlaceholderFloatingToolbar(editorView.state, editorView.dispatch);
+        expect(editorView.state.selection).toBeInstanceOf(Selection);
       });
     });
   });